move location to utils and use it for the array center
 remove old array center
